chore: add default watch excludes
